requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.3
Flask==3.1.0
Flask_JWT_Extended==4.7.1
Flask_Login==0.6.3
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
matplotlib==3.10.1
pandas==2.2.3
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
Requests==2.32.3
selenium==4.30.0
SQLAlchemy==2.0.38
Werkzeug==3.1.3
